requirements.txt
click>=6.0,<9.0
requests>2.0,<3.0
tabulate>=0.8
arrow>=0.10
PyYAML>=5.1
